Move request status
