+++ internal/lookout/ui/src/components/job-sets/JobSets.tsx
@@ -19,8 +19,8 @@ import { AutoSizer } from "react-virtualized"
 
 import RefreshButton from "../../RefreshButton"
 import { JobSetsView, isJobSetsView } from "../../containers/JobSetsContainer"
-import { RequestStatus } from "../../containers/JobsContainer"
 import { DurationStats, JobSet } from "../../services/JobService"
+import { RequestStatus } from "../../utils"
 import AutoRefreshToggle from "../AutoRefreshToggle"
 import DurationPlotsTable from "./DurationPlotsTable"
 import JobSetTable from "./JobSetTable"
