+++ internal/lookout/ui/src/components/jobs/CancelJobsModal.tsx
@@ -3,8 +3,8 @@ import React from "react"
 import { Backdrop, Fade, Modal, createStyles } from "@material-ui/core"
 import { makeStyles } from "@material-ui/core/styles"
 
-import { RequestStatus } from "../../containers/JobsContainer"
 import { CancelJobsResult, Job } from "../../services/JobService"
+import { RequestStatus } from "../../utils"
 import CancelJobs from "./CancelJobs"
 import CancelJobsOutcome from "./CancelJobsOutcome"
